agent/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
langgraph>=0.2
langchain-core>=0.3
langchain-openai>=0.2
google-auth>=2.30
httpx>=0.27
python-dotenv>=1.0
